MQTT moved to second window. Everything works now as intended.
